Feature/ecommerce client testing

+++ api/client/views/client/retrieve_update_destroy.py
@@ -1,18 +1,19 @@
 from knox.auth import TokenAuthentication
 from rest_framework import status
+from rest_framework.exceptions import PermissionDenied
 from rest_framework.generics import RetrieveUpdateDestroyAPIView, get_object_or_404
 from rest_framework.permissions import IsAuthenticated
 from rest_framework.response import Response
 
 from client.models.ecommerce_client import EcommerceClient
 from client.serializers.client import MeSerializer, EcommerceClientSerializer
-from common.views import HTTP_RETRIEVE_METHODS, HTTP_UPDATE_METHODS
 
 
 class EcommerceClientViewSet(RetrieveUpdateDestroyAPIView):
     authentication_classes = (TokenAuthentication,)
     permission_classes = (IsAuthenticated,)
     queryset = EcommerceClient.objects.all()
+    output_serializer_class = MeSerializer
 
     def get_serializer_class(self):
         return EcommerceClientSerializer
@@ -27,11 +28,15 @@ class EcommerceClientViewSet(RetrieveUpdateDestroyAPIView):
 
     def update(self, request, *args, **kwargs):
         partial = kwargs.pop('partial', False)
-        instance = self.get_object()
-        serializer = self.get_serializer(instance, data=request.data, partial=partial)
+        client = self.get_object()
+
+        if not client == request.user:
+            raise PermissionDenied()
+
+        serializer = self.get_serializer(client, data=request.data, partial=partial,
+                                         context={'user': request.user})
         serializer.is_valid(raise_exception=True)
         self.perform_update(serializer)
-
 
         # TODO: Update password system must be via a different endpoint auth/update_password/.
         #  Current password will have to be provided in this new endpoint along with the new password,
@@ -43,10 +48,14 @@ class EcommerceClientViewSet(RetrieveUpdateDestroyAPIView):
         # if password and not client.check_password(password):
         #     client.set_password(password)
 
-        return Response(instance.data, status=status.HTTP_200_OK)
+        return Response(self.output_serializer_class(client).data, status=status.HTTP_200_OK)
 
     def delete(self, request, *args, **kwargs):
         client = self.get_object()
+
+        if not client == request.user:
+            raise PermissionDenied()
+
         EcommerceClient.delete(client)
 
-        return Response(self.get_serializer(client).data, status=status.HTTP_200_OK)
+        return Response(self.output_serializer_class(client).data, status=status.HTTP_200_OK)
